requirements.txt
# Dependências diretas de produção
Flask==3.1.3
Flask-Babel==4.0.0
Babel==2.16.0
psycopg2==2.9.11
gunicorn==25.1.0
python-dotenv==1.2.2
Werkzeug==3.1.6
alembic==1.16.1
click==8.3.1
